pyproject.toml
[project]
name = "b5auth"
version = "0.1.0"
description = "B5 uchun test proyekt"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "argon2-cffi>=25.1.0",
    "cryptography>=45.0.5",
    "fastapi[all]>=0.115.13",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "pytz>=2025.2",
    "ruff>=0.12.0",
    "sqlalchemy>=2.0.41",
]

# RUFF configs
[tool.ruff]
target-version = "py312"
line-length = 88

# Enable rules
lint.select = ["E", "F", "I", "N", "UP", "B", "C90", "RUF"]

lint.ignore = [
    "E501",     # line too long
    "RUF001",   # ambiguous unicode character in code
    "RUF002",   # ambiguous unicode character in docstring
]  

# Paths to exclude
exclude = ["migrations", "__pycache__", ".venv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
